Compilable, but API-breaking changes.
This commit breaks the API. This commit compiles on Mac OS X, but may
not run correctly, as it is not tested.

The route class has been updated to use std::function objects for both
conditions ("SHOULD this connection be handled?") and handlers ("HOW
should this connection be handled?").

This means that the various condition* functor classes have been
removed, with the (request method && uri) and (request method and uri
regex) functionality moved into the overridden "on()" methods.

Updated the application class so that is now privately inherits from
boost::noncopyable. It also has protected inheritance from the router
class, allowing derived classes to delegate connections to appropriate
handlers.  This replaces the various connect() functions previous used.
It also includes the new setup_routes() function, which is used to
initialise the routes within the system (along with the clear() function
to remove all existing routes). The run() function is now no longer
virtual, and has been fleshed out further to use threads appropriately.

The connection class has been updated to use std::shared_ptr objects for
the request and response objects. This allows the connection class to be
returned and passed by value, while pointing to the same objects.  The
concept of "uninitialized" connections has also been added, with the
operator bool() function being used to return whether or not the
connection is initialized.

The gateway class has been updated. It is now an abstract base class,
and the CGI functionality has been moved to the cgi_gateway class.  Due
to changes in the connection class (that is, the fact that it can now be
returned and passed by value, and that an "uninitialized" connection can
be determined with the bool operator), the get_connection() method
returns a connection object.  It also now includes the is_threaded()
function so that other protocols (e.g., SCGI) will spawn a new thread
for each connection.  The gateway inherits from boost::noncopyable,
rather than specifying this restriction manually.

Added and updated source code documentation.