CGI-related updates.
Note that this commit breaks the API. It also compiles, although it will
likely not work as expected.

Refactored the request and response classes into cgi_request and
cgi_response.

The response_status enum class has also been moved to a member
enumeration, response::status_type.  This is keeping in standard with
how request methods were moved to request::method.