Updated connection and cgi_gateway.
Updated the connection class so that the request and response objects
are passed in to initialized connections via shared pointers. This
encourages the caller to wrap each "new" in a shared_ptr, to avoid
possible (although unlikely) memory leaks.

Also fixed a bug with cgi_gateway, where the connection counter was not
incremented (because it came after the return statement).

Updated todos.