Significant refactoring of session classes.
The session class now includes the session ID as a member variable.

Renamed session_storage -> session_backend, as this name is more
descriptive of the purpose.

Changed the meaning of the error session_file_directory_invalid to mean
that the directory is not a directory, as the constructor for the
session backend creates the directory if it does not exist.

The session constructor now repeatedly calls the backend has_session()
function to determine whether the session ID is unique.