Added UNIX compiler flags, and differentiated between it and APPLE compiler flags.
The application now only inherits privately from router, not from
boost::noncopyable. The reason for this is that router already inherits
privately from boost::noncopyable.

Changed the accessors for request and response objects from the
connection object to avoid an ambiguity warning on GCC. The router and
session calls to this accessor have also been updated.